# optimize cast op