### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] --> Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] --> APIs 

### Describe
<!-- Describe what this PR does --> return the input(x) when x.dtype is the same as attr(dtype)

### GPU Performance
V100, cuda10
|op | shape | input dtype|dtype | before |after| speed up|
|---|---|---|---|---|---|---|
|cast | [16, 1785] |bool| bool|0.00150 ms|0.000026 ms  | 57x|